Move getArticleIds to mw-api, minor fixes
