+++ src/util/builders/url/basic.director.ts
@@ -3,7 +3,7 @@ import urlBuilder from './url.builder.js'
 type DownloaderBaseUrlConditions = Array<{ condition: boolean; value: string }>
 
 /**
- * Build base URL for specific wiki. Expamle of the output: 'https://en.wikipedia.org' or 'https://de.wikivoyage.org
+ * Build base URL for specific wiki. Example of the output: 'https://en.wikipedia.org' or 'https://de.wikivoyage.org
  */
 class BasicURLDirector {
   buildMediawikiBaseURL(domain: string) {
