+++ test/unit/urlRewriting.test.ts
@@ -2,7 +2,7 @@ import { startRedis, stopRedis, redisStore } from './bootstrap.js'
 import domino from 'domino'
 import { rewriteUrl } from '../../src/util/rewriteUrls.js'
 import { makeLink, setupScrapeClasses } from '../util.js'
-import { getArticleIds } from '../../src/util/redirects.js'
+import { getArticleIds } from '../../src/util/mw-api.js'
 import { saveArticles } from '../../src/util/saveArticles.js'
 import { ZimArticle } from '@openzim/libzim'
 import { mwRetToArticleDetail } from '../../src/util/index.js'
